🚚 [RUMF-1084] move synthetics context getter in its own module
