+++ packages/rum-core/src/tools/syntheticsContext.ts
@@ -1,0 +1,22 @@
+import { getCookie } from '@datadog/browser-core'
+
+export const SYNTHETICS_TEST_ID_COOKIE_NAME = 'datadog-synthetics-public-id'
+export const SYNTHETICS_RESULT_ID_COOKIE_NAME = 'datadog-synthetics-result-id'
+
+export interface BrowserWindow extends Window {
+  _DATADOG_SYNTHETICS_PUBLIC_ID?: string
+  _DATADOG_SYNTHETICS_RESULT_ID?: string
+}
+
+export function getSyntheticsContext() {
+  const testId = (window as BrowserWindow)._DATADOG_SYNTHETICS_PUBLIC_ID || getCookie(SYNTHETICS_TEST_ID_COOKIE_NAME)
+  const resultId =
+    (window as BrowserWindow)._DATADOG_SYNTHETICS_RESULT_ID || getCookie(SYNTHETICS_RESULT_ID_COOKIE_NAME)
+
+  if (typeof testId === 'string' && typeof resultId === 'string') {
+    return {
+      test_id: testId,
+      result_id: resultId,
+    }
+  }
+}
